Numbers length had limited to int32 and float32

+++ leibniz.java
@@ -1,10 +1,10 @@
 public class leibniz {
     public static void main(String[] args) {
-        double pi = 4;
+        float pi = 4.0f;
         int sign = -1;
         long start_at = System.currentTimeMillis();
         for (int i = 3; i < 2147483647; i += 2) {
-            pi = pi + (double) sign * 4 / i;
+            pi = pi + (float) sign * 4 / i;
             sign *= -1;
         }
         System.out.println(pi + " consume " + (System.currentTimeMillis() - start_at) + " msec");
